feat: give typed arrays a shared %TypedArray%.prototype with the iteration methods
A typed array in sai inherits nothing but Object.prototype, so none of the
iteration methods ES6 22.2.3 puts on %TypedArray%.prototype are reachable:
new Int8Array([1,2,3]).map(f) is a TypeError, and the same holds for join,
forEach, every, some, filter, reduce and reduceRight. Reaching them today
means Array.prototype.map.call(view, f), which also hands back a plain Array
instead of a typed array.

There are nine concrete typed array classes and one common base,
ArrayBufferView, but saigen only ever looks at the class it is processing:
a @Function on ArrayBufferView would land on ArrayBufferView$Prototype, and
nothing installed that class. Writing the methods out nine times is the only
alternative, and that is around 1500 lines of delegation.

So install it instead. ArrayBufferView now carries the @Function methods and
saigen generates ArrayBufferView$Prototype from them; Global instantiates that
class once per Global and splices the instance in as the [[Prototype]] of each
of the nine concrete prototypes, which is the shape ES6 describes. The nine
getBuiltinXxxArray() methods go through one initTypedArrayConstructor() helper
that does the splice, so the wiring exists in a single place. Own properties of
the concrete prototypes are untouched, and the new methods are non-enumerable,
so nothing that enumerates a typed array or its prototype sees a difference.

join, forEach, every, some, reduce and reduceRight are pure delegation to the
NativeArray implementations, which already work over any array-like. map and
filter cannot be: they have to return a typed array of the receiver's own type,
so they build the result through the receiver's Factory. filter collects the
kept elements first because a typed array's length is fixed at construction.

Every method starts at ValidateTypedArray (ES6 22.2.3.5.1), so a null or
undefined receiver - or a plain Array - is a TypeError before any element is
read, rather than something that half-works.

Writing an Object into a Uint8ClampedArray turned out to be a StackOverflowError:
Uint8ClampedArrayData.set(int, double, boolean) called set(index, rint(value),
strict), and rint returns a double, so the call resolved back to itself. Nothing
reached it before, because stores from script go through the SET_ELEM method
handle, but map and filter do. Casting rint's result to int picks the
set(int, int, boolean) overload that was meant.

+++ src/main/java/org/codelibs/sai/internal/objects/NativeUint8ClampedArray.java
@@ -201,7 +201,8 @@ public final class NativeUint8ClampedArray extends ArrayBufferView {
 
         @Override
         public ArrayData set(final int index, final double value, final boolean strict) {
-            return set(index, rint(value), strict);
+            // rint() returns a double, so without the cast this would call itself forever
+            return set(index, (int) rint(value), strict);
         }
 
         private static double rint(final double rint) {
